Ensure s3 deploys have correct content types
Twitter cards only parse pages with an explicit content-type: text/html
declaration. S3-based deploys need this set through the aws api.

This adds a sample _deploy.yml with the correct header section, so that
the octopress deploy command will set the content-type headers
correctly. The user can use this as a template for other file-types, as
needed.

The end user, when configuring the template for eventual use, will have
to merge this section with their own s3 deploy setup.