Add image alts to site config and social cards
Image alt text is important for accessible site design, and both Twitter
and Facebook consider it required for their social card integration.

This adds an alt text configuration everywhere an image is configured.
It then adds the appropriate image alt meta tags for Twitter and
Facebook social cards.

It does not yet add image alt text to page and post headers: that is a
future TODO.