Update to Ruby 2.5.1 and remove jekyll-assets
Updating the Ruby version to ensure we stay up-to-date and to fix some
gem errors that required the personalized version of the blog to use
`bundle exec jekyll`

Jekyll assets is not longer needed with the final versions of bootstrap
and was leaving build artifacts in the built _site
(/assets/*manifest.json). It has now been removed.