Add explicit timezone and permalink definition
Add an explicit timezone to the site configuration. Because the sample
post date is set to exactly midnight on 2017-01-01 in UTC but my test
builds were done on a system in America/New_York, the sample builds kept
treating the sample post as written on 2016-12-31.

Add an explicit permalink, and change it to exclude the day of the
month, and output pretty urls. This results in post urls looking like
`/2017/01/01/sample-post/` rather than
`/2017/01/01/01/sample-post.html`, which IMO is a bit too verbose.