Expand "Powered by Jekyll" section of footer
The "Powered by Jekyll" of the site footer didn't have an actual link to
Jekyll, and made no mention of this particular theme or it's author.

While site maintainers are of course free to remove this section if they
desire, including it by default can give exposure to the Jekyll theme
project so I'm including it.