Refactor sharing include and add to resume
The inclusion of the social_sharing include didn't match other
section-level includes in the placement of the section tag,leading to
inconsistent markup in the templates. Furthermore, sharing links were
only added to the post template and not to pages or the resume.

This refactors the sharing include so now all section-level includes are
used in the form:
```
<section clas="...">
{% include section_name.html %}
</section>
```

Furthermore, the sharing section is added to the resume layout, and to
the page layout if the page frontmatter sets `sharing: true`.