Extract header to _includes
There was a lot of duplication in creating page and post headers that
could easily be extracted, and some verbose use of `if` blocks that
could easily be replaced by either `default:` filters or `assign`
variables and creative css.

This will also make it much easier to eventually add header image
attribution blocks, since it only needs to be added to one template.