Resume fixes: Hide btns on print, multi-size jobs
There were a couple of small problems with resume template needing
hotfixes.

* Printed resume shows sharing buttons and the horizontal rule
separating them from the rest of the resume. This was fixed by adding
both to a div, then hiding the div on print media.

* Resume jobs of widths besides 6 didn't get a padding-bottom. This was
fixed by adding an additional .resume-job class to the jobs, and using
that class to select for the padding-bottom.

* Resume download link should use white text. This was, in fact, fixed
in the commit just prior extracting the header template, but I'm
mentioning it here because these two commits were written together.