Add pre-post asides
In order to call out special introductory sections that aren't part of
the post proper, the correct lexical syntax is to use an aside.
Furthermore, this content is generally not desired in excerpts or
syndicated (e.g. atom/rss) feeds, so including asides as part of the
post content is incorrect.

This feature adds support for YAML frontmatter specifying one or more
`pre-aside` blocks. Each of them supprots:

* `class`, one of the four general-purpose bootstrap "info", "success",
  "danger", or "warning". This is easy to extend in the asides.scss
* heading: The text of the aside heading
* text: The text of the aside.

The nameing as `pre-aside` assumes a `post-aside` feature can easily be
added in the future.