Hotfix: Paragraphs nested in divs treated as leads
The styling to treat the first paragraph of a post as a lead paragraph
was incorrectly nesting to treat the first paragraph of *any* div
withing a post as a lead paragraph. This lead to odd behavior such as
not bolding text within captions.

Solved by adding a class to the nearest wrapping div of the post content
(called .main-content), and selecting the lead paragraph as
`>p:first-child` instead of `p:first-child`