Add a code of conduct to the site
This site template encourages the use of Disqus comments on posts (and
other sections).

It is both Disqus' recommendation and an overall matter of good form to
include a Code of Conduct that applies to the comments section, as well
as any other side channel communications (e.g. Twitter threats, Slack,
etc).

Note that this is **not** the CODE_OF_CONDUCT for *contributors* the the
project itself: that is a (soon) planned addition.

I named it `conduct.md` with an explicit permalink so an eventual
CODE_OF_CONDUCT.md can be added that won't conflict on operating systems
without case sensitive file systems.