Hotfix: Spacing and hrs below tags/archive asides
The spacing below the archive page contents aside was too small on
mobile devices. Furthermore unlike the tags page contents aside, there
was no horizontal rule on mobile.

This adjusts the spacing to be consistent between the two and ensures
the horizontal rules disappear at the same breakpoint that the aside
moves to the side of the display.