Hotfix: Tweaks to positioning of image figures
Image figures on mobile were properly scaling to fill the device, but
improperly adding a margin to just one size and appearing off center.

This commit ensures that the float left/right are only applied to small+
devices (eg not phones). It also tweaks some of the sizing classes for
more sensible scaling options.