Move navbar scripts to scripts partial
Removing the extra jquery script load from the head section broke the
navbar active item selection script, because it was being run before the
scripts at the end of the page.

This moves that selection script to script section.

Also, it removes an extra level of indentation in the same file.