Fix category stripe on index cards with images
The category stripe have had odd edges for some time (due to the
different width/color between top & side borders). Recently, I noticed
that index cards with preview images had the stripe in the middle,
and not on top.

This commit adds an hr element to the index cards specifically to be
used as the category stripe, and styles/places it (including rounding
the corner without the tab) appropriately.