Downgrade to Ruby 2.3.8 to fix dependency issues
In order to deploy to AWS using octopress, the `aws-sdk-v1` gem is
required. This gem, in turn, requires the `json` gem at version 1.x.
However, starting with Ruby 2.4, version 2.x of `json` is included as a
default gem. This resulted in dependency errors throughout the project.

This change downgrades the Ruby version to 2.3.8 in order to allow
octopress to continue deploying to AWS.