Upgrade to Ruby 2.7.2 and all bundle dependencies
With Ruby 2.3.8 getting awfully old --- and difficult to install on
modern operating systems with its dependency on libssl-1.0 --- I've
upgraded the Ruby version to 2.7.2.

Upgrading the Ruby version lead to several deprecation warnings in
jekyll and rb-inotify, so I also updated the entire bundle. Doing so
required adding kramdown-parser-gfm to the jekyll_plugins group, so
that's now part of the bundle as well.

I'm uncertain if the project will continue to work on Ruby 3.0, which is
going to be released later this month. We'll see once the 25th comes
along :)