Install s3_website gem to replace octopress deploy
Octopress's built-in deploy command no longer works with modern versions
of Ruby: Octopress itself was last updated in 2015, so this is not
unexpected.

The Jekyll docs suggest using `s3_website`, a continuation of `jekyll_s3`
with support for arbitrary site builders.

This commit adds the s3_website gem and a configuration file with some
sensible defaults for a jekyll blog. It incorporates dotenv support, so
the end user has to supply their AWS credentials in a `.env` file (which
is *not* kept in version control).