Update bundle to fix security vulnerabilities
While technically this is an offline application, and these security
vulnerabilities won't affect a deployed *static* site, there's no harm
in updating the bundled gems so long as we don't break compatibility
with jekyll/octopress et al.

The `json` gem is being kept back because the only available fix is in
the 2.x branch, and we need the 1.x branch for our dependencies, at
least without some major work.