Related posts
Added a *related posts* section to the theme. This section is either a
full-width banner or a 12-column card (configurable in `_config.yml`)
between the *social links* and comments *comments* sections.

The posts show up as a bootstrap4 carousel, with indicators and
forward/backward controls. The images for the carousel are decided by
the header images, in the following order, reflecting the changes made
to the post headers in the previous commit:
  * the post, if available
  * the post's category, if the post has no specific header image
  * the site's default header image, if neither of the previous are
    found

The `classifier-reborn` gem is added to enable the `--lsi` argument with
Jekyll's build/serve commands. This should make selecting related posts
better, once a site has a large enough pool of posts to choose from.

Two configuration options are supported in `_config.yml` for this
feature:
  * `related_count: 2` is the number of related post to choose.
  * `related_wide: true` is, as described, whether two display the section
    as a full-width banner (or else a 12-column card).

Also added two new example draft posts, to help show off the related
posts section. These posts will only be included if the site is
previewed/build using the `--drafts` argument, e.g.

While working with drafts, I updated the (mostly empty) `drafts` template
to be more in line with the `posts` template. The existing templates and
example post were tweaked to make explicit the rules for selecting only
**one** category.

```
jekyll serve --drafts --unpublished
```