Make styled alerts/asides available within posts
We had previously implemented pre-asides, which were defined in a post
front-matter to appear before the post content as styled alert boxes,
complete with icons. Here we extend that to allow blocks within post
content to be similarly styled.

Support for making one of the asides (or pre-asides) on the post sticky
has also been added. The sticky aside will stay on top of the screen (on
medium+ devices) until the main content is done. It will move off to
make room for the sharing, related posts, and comments sections.

The example post is expanded a bit to demonstrate the use of these new
features. The templates are updated to make their warning pre-asides
sticky.

A quick fix to change screen-reader only headings to `d-none` is made
here as well.