Refactor and lint dark.css to scss
Refactored toe `dark.css` file to `dark.scss`, and made use of both
nesting and scss variables to make the code a bit cleaner. The use of
variables in particular will make playing with the color scheme a lot
easier.

Reordered the imports in `site.css` import all the code-related blocks
together.

The `xcode.css` can probably be changed to a `light.scss` import in an
upcoming commit.