Add example of console lexer to example posts
Added a thorough example of the `console` type code fences (and its
alias), demonstrating how the various `console`-specific classes in
`dark.scss` affect it's output.

Included is a listing of its various options. One of those options is a
custom option on a yet-to-be released fork of `rouge` --- the lexing of
root prompts separate from regular user prompts --- which may
eventually be added to this theme's Gemfile

The `dark.scss` theme was further tweaked and refactored using the new
live console example to make everything look good.

The front-matter for `example-2.md` was updated with tags that better
indicate it's useful content, and pre-asides that better explain both
it's use as an example post and the unreleased nature of the `rprompt`
settings.