Run sass-lint and fix violations
Run sass-lint on all the sass and fix violations as found.,

I made use of line-disabling on a few places. In particular, I'm keeping
a few class+pseudo selectors un-nested when not using the selector by
itself.

There's a bit more extraction and refactoring to do from here; in
particular by extracting all variables to the `_variables.scss`. We also
need to get the `.dark` class working with `render_code` blocks.