Update rouge to my feature.root-prompt branch
Changed the `rouge` gem to point to the `feature.root-prompt` branch
under my github fork `etothepiipower/rouge`. Supporting root prompts as
separately-styled from user prompts is part of the plans for 1.0, so
this is within scop, unless/until the feature is added to the main
`rouge` release.

Updated some of the example posts to reflect this change.

Added a specific version of bootstrap to the Gemfile to prevent bundler
from trying to update to 5.x, which would require massive rewrites to
the theme.