Update .sticky class to mobile-first design
Updated the `.sticky` and `.sticky-xs` classes to `.sticky` and
`.sticky-sm` classes, with the base-line `.sticky` class being sticky on
all screen sizes. This better follows the mobile-first design principle,
leaving the move explicit class to be used to limit larger sticky asides
to appear only on large-enough screens.

Updated all sample posts and templates to be consistent in using
`.sticky`

Updated the `toggle-sticky.js` helper function for the new classes. It
now toggles the `#sticky` id explicitly, as our design principle calls
for only one aside to be set as sticky at any time, and this always in
the pre-asides.