Fix: Category stripe on index cards & misplaced hr
The category stripe on index cards was occasionally displayed
incorrectly, with a white line inside it. This caused because it was
rendered as a horizontal rule with top & bottom borders and a
*supposedly* empty content, which would sometimes not be rendered
entirely empty. The line is now the top-border of the card body's
header.

The horizontal rule, meanwhile, was in a semantically inappropriate
place and didn't look correct on text-only browsers or sound correct
with screen readers; it separated the category tab from the
corresponding card. It as been moved to the end of the logical card (and
the beginning of the next one), making for a more semantically sensible
layout.