Fix: related-posts image doesn't cover card on sm
Fixed a glitch where the images in the related-posts image carousel
didn't fully fill the card on small displays.

Fixed this by setting the `object-fit: cover` property, with a `width:
100%` and `height: 30em`. This height was chosen because the overall
carousel height was being set to 20em, but we needed to go a bit larger
to fill the card. 30em fills the card at the narrowest tested widths as
well as full-screen 4k; which should be fine for any real-world device/browser widths.