A11y: Add aria attributes and roles to `navbar`
Added `aria-label` and `role` attributes to `_includes/navbar`. This
should improve accessibility with screen readers.

Added `aria-hidden` attributes to `author_popover` and
`header_attributes` . The author popover provides repetitive (but
decorative) information that can already be found in the sidebar (on
posts) or by navigating to the author's bio (on index pages). The author
name and link will still be included with screen readers, so this is a
net improvement for screen-reader accessibility.

As for the header attribution; as the header images themselves are
purely decorative, its relevance *probably* isn't balanced by the length
and detraction from the actual content that including the attributions
would entail. However, this is debatable, and we might revert to keeping
the attributions in screen readers as an `aside`, i.e.
`complementaryinfo`.