Fix: Category stripe on index pages w/ header
This is a fix for a regression caused by removing the horizontal rule
in Commit 7d24dfc

On index pages, post excerpt cards with header images weren't displaying
the stripe in the correct place. Instead of below tab, the stripe was
placed between the image and the card body.

The fix adds a second rule to the style to add the stripe on top of the
card-img-top, along with an additional rule to *unset* the stripe on a
card-body immediately following a card-img-top