Tweak: tag-index sidebar too narrow on md devices
The sidebar index on the tag-index page was too narrow on medium
devices. The `.col-3-md` simply couldn't fit many tags on a single line
in the list-group, leading to inconsistent and ugly aesthetics.

This change updates the sidebar to a `.col-4-md.col-3-lg`, and the
corresponding main content section to `.col-8-md.col-9-lg`, which gives
the sidebar more horizontal space that should fit any reasonably long
tag.