Fix: Archive loop inefficient and shows empty years
The old archive page loop was inefficient. It used to:

* Loop over *every* year from the first post to the last, even if there
  were no posts in that year
* For each year, create a list group (which might be empty)
* Then loop over *every* post in the whole collection, only listing the
  posts for the given year.

This is an O(n*N), with `n` the number of posts and `N` the number of
the years the blog has been active, including potentially inactive
years.

With this new refactor, the archive page runs a `group_by` filter over
the collection of posts, then only for years for which posts exist,
loops through *only those* posts.

This results in both a performance gain (of questionable utility in a
statically generated site, but still good), and eliminating empty years
from the archive page. This latter benefit may be of particular note for
bloggers who took a break.

It may also be useful for anyone wishing to
extend the Archives to a Year-Month format.